Add tests for ToMarkdown service
