+++ src/cloudflare/internal/test/to-markdown/python-to-markdown-api-test.wd-test
@@ -1,0 +1,36 @@
+using Workerd = import "/workerd/workerd.capnp";
+
+const unitTests :Workerd.Config = (
+  services = [
+    ( name = "to-markdown-api-test",
+      worker = (
+        modules = [
+          (name = "worker.py", pythonModule = embed "to-markdown-api-test.py")
+        ],
+        compatibilityDate = "2023-01-15",
+        compatibilityFlags = ["nodejs_compat", %PYTHON_FEATURE_FLAGS],
+        bindings = [
+        (
+          name = "ai",
+          wrapped = (
+            moduleName = "cloudflare-internal:ai-api",
+            innerBindings = [(
+              name = "fetcher",
+              service = "to-markdown-mock"
+            )],
+          )
+        )
+        ],
+      )
+    ),
+    ( name = "to-markdown-mock",
+      worker = (
+        compatibilityDate = "2023-01-15",
+        compatibilityFlags = ["experimental", "nodejs_compat"],
+        modules = [
+          (name = "worker", esModule = embed "to-markdown-mock.js")
+        ],
+      )
+    )
+  ]
+);
